Update strat command to accept town hall input
