+++ Dockerfile
@@ -3,4 +3,5 @@ WORKDIR /app
 COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 COPY bot.py .
+COPY cocbot ./cocbot
 CMD ["python", "bot.py"]
