+++ requirements.txt
@@ -1,3 +1,4 @@
 discord.py>=2.5
 coc.py>=3.9
 boto3>=1.34
+openai>=1.3
